requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0
pillow>=10.0.0
tensorflow>=2.15.0
imagehash>=4.3.0
jupyter>=1.0.0
notebook>=7.0.0
streamlit>=1.30.0